Extension Features + Peer Feedback

+++ projectaria/src/components/StudySessionPage/SessionHistoryCard.jsx
@@ -15,9 +15,13 @@ import {
 import { Cancel } from "@mui/icons-material";
 // supabase imports
 import { supabase } from "../../supabase";
+// component imports
+import { buttonColors, possibleFontColors } from "../themes";
 
 export default function SessionHistoryCard({
   userID,
+  email,
+  presentTheme,
   retrieveSessionHistory,
   sessionHistoryArray,
   setSessionHistoryCardOpen,
@@ -25,11 +29,15 @@ export default function SessionHistoryCard({
   /* Component functionality */
   async function clearSessionHistory(e) {
     e.preventDefault();
-    const { error } = await supabase
-      .from("studysessions")
-      .delete()
-      .eq("user_id", userID);
-
+    await supabase.from("studysessions").delete().eq("user_id", userID);
+    await supabase
+      .from("stats")
+      .upsert({
+        user_email: email,
+        failStreak: 0,
+        successStreak: 0,
+      })
+      .select();
     retrieveSessionHistory();
   }
 
@@ -39,53 +47,49 @@ export default function SessionHistoryCard({
   }
 
   return (
-    <Container>
+    <Container sx={{ backgroundColor: "transparent" }}>
       <Card
         id="session-history-container"
         sx={{
-          height: "445px",
+          // height: "445px",
           overflow: "auto",
           padding: 2,
         }}
+        elevation={0}
       >
         <Button
           id="clear-session-history-button"
-          sx={{ float: "right", fontFamily: "inherit", marginBottom: 2 }}
-          color="secondary"
+          sx={{
+            float: "right",
+            fontFamily: "inherit",
+            marginBottom: 2,
+          }}
+          color={buttonColors[presentTheme]}
           disabled={sessionHistoryArray.length === 0}
           onClick={clearSessionHistory}
-          variant="outlined"
+          variant="contained"
         >
           Clear History
         </Button>
         <IconButton
           id="close-session-history-button"
           onClick={closeSessionHistory}
-          sx={{ float: "left", color: "black" }}
+          sx={{ float: "left", color: possibleFontColors[presentTheme] }}
         >
           <Cancel fontSize="inherit" />
         </IconButton>
         <br />
         <TableContainer component={Paper}>
-          <Table id="session-history-table" sx={{ backgroundColor: "#DC9A7F" }}>
+          <Table id="session-history-table">
             <TableHead id="session-history-table-header">
               <TableRow>
-                <TableCell
-                  sx={{ fontFamily: "Ubuntu", color: "white" }}
-                  align="left"
-                >
+                <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                   Timestamp
                 </TableCell>
-                <TableCell
-                  sx={{ fontFamily: "Ubuntu", color: "white" }}
-                  align="left"
-                >
+                <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                   Duration
                 </TableCell>
-                <TableCell
-                  sx={{ fontFamily: "Ubuntu", color: "white" }}
-                  align="left"
-                >
+                <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                   Status
                 </TableCell>
               </TableRow>
@@ -96,16 +100,10 @@ export default function SessionHistoryCard({
                   key={index}
                   sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
                 >
-                  <TableCell
-                    sx={{ fontFamily: "Ubuntu", color: "white" }}
-                    align="left"
-                  >
+                  <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                     <p>{session.created_at}</p>
                   </TableCell>
-                  <TableCell
-                    sx={{ fontFamily: "Ubuntu", color: "white" }}
-                    align="left"
-                  >
+                  <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                     <p>
                       {session.duration.hours === ""
                         ? 0
@@ -113,10 +111,7 @@ export default function SessionHistoryCard({
                       h {session.duration.minutes}m
                     </p>
                   </TableCell>
-                  <TableCell
-                    sx={{ fontFamily: "Ubuntu", color: "white" }}
-                    align="left"
-                  >
+                  <TableCell sx={{ fontFamily: "Ubuntu" }} align="left">
                     {session.completed ? (
                       <p style={{ color: "#32CD32" }}>Success</p>
                     ) : (
